Add files via upload
Objectives

Use Excel to analyze the sales based on various product categories.

Description

The dataset in file E-Commerce Dashboard dataset.xlsx contains sales data for different product categories. The following are the features in the dataset.

Problem Statement and Motivation

Design a sales dashboard that analyzes the sales based on various product 
categories. The company wants to add user control for product categories 
so that users can select a category and can see the trend month-wise and 
product-wise accordingly